refactor(react-grid): Move layout templates to Grid (#107)
BREAKING CHANGE:

We moved layout templates from the TableView plugin to the Grid component. The Grid component with the predefined plugins is now available in the "@devexpress/dx-react-grid-bootstrap3" package. If you want to define custom layout templates, the Grid component without predefined templates is still available in the "devexpress/dx-react-grid" package.

The following code:

```jsx
import {
  Grid
} from '@devexpress/dx-react-grid';
```

should be replaced with:

```jsx
import {
  Grid
} from '@devexpress/dx-react-grid-bootstrap3';
```