refactor(react-grid): replace render functions with components in GroupingPanel (#540)
BREAKING CHANGES:

The GroupingPanel plugin's `groupPanelTemplate`, and `groupPanelItemTemplate` properties have been replaced with  `containerComponent`, and `itemComponent` ones, which accept components instead of render functions. Find more details here: #496 

`containerComponent` takes on the `children` property instead of all arguments passed to the `groupPanelTemplate` function.

The `onSort`, and `onGroup` properties passed to `itemComponent` are used instead of the `changeSortingDirection`, and `groupByColumn` arguments passed to the `groupPanelItemTemplate` function. The `item`, and `draft` properties are no longer available, use the `item` property instead.